wip: split mapper into classes
